Remove spurious files and add them to .gitignore
