+++ .gitignore
@@ -5,3 +5,4 @@ my_main.c
 my_printf.c
 print_char.c
 print_int.c
+hol_main.c
