--- test-mains/README.md
@@ -1,1 +1,0 @@
-Custom main files.
